Add ability to only allow moves which move king out of check
A deep copy of the chess board is made and the player move is trailed.
if the resulting move moves the players king out of check then the main
board is updated if not the players piece is returned to its original
position and the player must make another move.

There was a bug when implementing this. I was original using memcpy to
copy the board however since the Board struct contains pointers I needed
a deep copy to be made, to avoid changes involving the pointers to
not affect the main board.