Add freeBoard function to board.c
freeBoard function cleans up the dynamic memory for Board structs

Bug: There was a bug found in endDragOperation in main.c. Both the main
board and the test board where updated using a pointer to the same
Piece struct in memory. Therefore when the memory was cleaned for the
test board the new piece (copy of the drag piece) was being freed from
both the test board and the main board.

Solution: make a separtate copy of the piece variable (testPiece) and
use that to update the testBoard. Now the test board and the main board
are complete deep copies of one another and changes to one should not
effect the other.