Update getEnPassantCells function in board.c
getEnPassantCells now checks if the player will be put into check by
capturing the EnPassant pawn. If yes the cell is not added to the linked
list and the player cannot captue that pawn in EnPassant.

This logic was previously in the endDragOperation function in main.c
and has been moved.